requirements.txt
selenium>=4.18.1
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0